+++ app/db/models/__init__.py
@@ -1,10 +1,9 @@
 from app.db import db
 from app.db.models.blog import BlogPost
-from app.db.models.contracts import Contracts
 from app.db.models.customer import User, UserRoles, Role, Company, CompanyMembers
 from app.db.models.forum import ForumComment, ForumPost, Thread, ThreadRoles
-from app.db.models.customer import Order, ShippingAddress, Customer
 from app.db.models.product import Product
+from app.db.models.orders import Contracts, Order, ShippingAddress, Customer
 from app.db.models.ticket import TicketComment, TicketForum, Resolution
 from app.db.models.calendar import Event
 from app.db.models.home import Home
