+++ app/db/models/customer.py
@@ -1,5 +1,4 @@
 from flask_login import UserMixin
-from sqlalchemy import func, schema  # , ForeignKey
 from datetime import datetime
 from app.db import db
 from app.util.uuid import id as ID
@@ -19,6 +18,8 @@ class User(UserMixin, db.Model):
     roles = db.relationship('Role', secondary='user_roles')
     posts = db.relationship('ForumPost')
     comments = db.relationship('ForumComment')
+    order = db.relationship('Order')
+    customer = db.relationship('Customer')
     companies = db.relationship('Company', secondary='company_members')
     tickets = db.relationship('TicketForum')
     ticket_comments = db.relationship('TicketComment')
@@ -50,51 +51,6 @@ class UserRoles(db.Model):
                                                    ondelete='CASCADE'))
 
 
-class Order(db.Model):
-    __tablename__ = 'order'
-    id = db.Column(db.String(36), default=ID, primary_key=True)
-    order_date = db.Column(db.DateTime(255), nullable=True)
-    product_id = db.Column(db.String(36), db.ForeignKey('product.id',
-                                                      ondelete='CASCADE'))
-    # service_date = db.Column(db.DateTime(255), nullable=True)
-    # expiration_date = db.Column(db.DateTime(255), nullable=True)
-    customer_id = db.Column(db.String(36), db.ForeignKey('customer.id',
-                                                       ondelete='CASCADE'))
-
-
-class ShippingAddress(db.Model):
-    __tablename__ = 'shipping_address'
-    id = db.Column(db.String(36), default=ID, primary_key=True)
-    address1 = db.Column(db.String(50))
-    address2 = db.Column(db.String(50))
-    city = db.Column(db.String(50))
-    state = db.Column(db.String(50))
-    country = db.Column(db.String(50))
-    phone = db.Column(db.String(50), unique=True)
-    email = db.Column(db.String(50), unique=True)
-
-
-class Customer(db.Model):
-    # Customer can be a user or a company, it can also be both
-    __tablename__ = "customer"
-    __table_args__ = (
-        schema.CheckConstraint('NOT(user_id IS NULL AND company_id IS NULL)'),
-    )
-    id = db.Column(db.String(36), default=ID, primary_key=True)
-    user_id = db.Column(db.String(36),
-                        db.ForeignKey('user.id', ondelete='CASCADE'),
-                        nullable=False)
-    company_id = db.Column(db.String(36),
-                           db.ForeignKey('company.id', ondelete='CASCADE'),
-                           nullable=False)
-    orders = db.ForeignKey('order.id', ondelete='CASCADE')
-    # ticket_id= db.Column(db.Integer(),
-    #                       db.ForeignKey('ticket.id', ondelete='CASCADE'))
-    date = db.Column(db.DateTime(timezone=True), server_default=func.now())
-    service_date = db.Column(db.DateTime(255), nullable=True)
-    expiration_date = db.Column(db.DateTime(255), nullable=True)
-
-
 class Company(db.Model):
     __tablename__ = "company"
     id = db.Column(db.String(36), default=ID, primary_key=True)
@@ -108,7 +64,7 @@ class Company(db.Model):
     address1 = db.Column(db.String(50))
     address2 = db.Column(db.String(50))
     members = db.relationship('User', secondary='company_members',overlaps="companies")
-    customer = db.relationship('User', secondary='customer')
+    customer = db.relationship('Customer')
 
 
 class CompanyMembers(db.Model):
